refactor: optimize get pod/service netowrk cidr logic
