+++ pkg/util/cidr_test.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"net"
+	"reflect"
 	"testing"
 	"time"
 
@@ -69,7 +71,7 @@ func TestByCreateSvc(t *testing.T) {
 
 func TestElegant(t *testing.T) {
 	before()
-	elegant, err := GetCIDRElegant(context.Background(), clientset, restconfig, namespace)
+	elegant, err := GetCIDR(context.Background(), clientset, restconfig, namespace)
 	if err != nil {
 		t.Error(err)
 	}
@@ -126,4 +128,182 @@ func TestPatch(t *testing.T) {
 		panic(err)
 	}
 	fmt.Println(string(bytes))
+}
+
+func TestRemoveCIDRsContainingIPs(t *testing.T) {
+	tests := []struct {
+		name          string
+		cidrStrings   []string
+		ipStrings     []string
+		expectedCIDRs []string
+		expectPanic   bool
+	}{
+		{
+			name: "Normal case - some overlaps",
+			cidrStrings: []string{
+				"10.140.45.0/24", "10.140.44.0/24", "10.31.0.0/24", "10.31.1.0/24", "10.31.2.0/24", "10.31.3.0/24", "10.140.47.0/24", "10.140.46.0/24",
+			},
+			ipStrings: []string{
+				"10.140.45.1", "10.140.46.220", "10.140.45.180", "10.140.45.152",
+				"10.140.46.183", "10.140.45.52", "10.140.47.148", "10.140.46.214",
+			},
+			expectedCIDRs: []string{
+				"10.140.44.0/24", "10.31.0.0/24", "10.31.1.0/24", "10.31.2.0/24", "10.31.3.0/24",
+			},
+			expectPanic: false,
+		},
+		{
+			name:        "Empty CIDR list",
+			cidrStrings: []string{},
+			ipStrings: []string{
+				"10.140.45.1",
+			},
+			expectedCIDRs: []string{},
+			expectPanic:   false,
+		},
+		{
+			name: "Empty IP list",
+			cidrStrings: []string{
+				"10.140.45.0/24", "10.140.44.0/24",
+			},
+			ipStrings: []string{},
+			expectedCIDRs: []string{
+				"10.140.45.0/24", "10.140.44.0/24",
+			},
+			expectPanic: false,
+		},
+		{
+			name: "All CIDRs removed",
+			cidrStrings: []string{
+				"10.140.45.0/24", "10.140.46.0/24",
+			},
+			ipStrings: []string{
+				"10.140.45.1", "10.140.46.220",
+			},
+			expectedCIDRs: []string{},
+			expectPanic:   false,
+		},
+		{
+			name: "Overlapping CIDRs",
+			cidrStrings: []string{
+				"10.140.45.0/24", "10.140.45.0/25", "10.140.45.128/25",
+			},
+			ipStrings: []string{
+				"10.140.45.1", "10.140.45.129",
+			},
+			expectedCIDRs: []string{},
+			expectPanic:   false,
+		},
+		{
+			name: "Invalid CIDR format",
+			cidrStrings: []string{
+				"10.140.45.0/24", "invalid-cidr",
+			},
+			ipStrings: []string{
+				"10.140.45.1",
+			},
+			expectedCIDRs: nil, // Panic expected
+			expectPanic:   true,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			defer func() {
+				if r := recover(); r != nil {
+					if !test.expectPanic {
+						t.Errorf("unexpected panic: %v", r)
+					}
+				} else if test.expectPanic {
+					t.Errorf("expected panic but got none")
+				}
+			}()
+
+			var cidrs []*net.IPNet
+			for _, cidr := range test.cidrStrings {
+				_, ipNet, err := net.ParseCIDR(cidr)
+				if err != nil {
+					if test.expectPanic {
+						panic(err)
+					}
+					t.Fatalf("failed to parse CIDR %s: %v", cidr, err)
+				}
+				cidrs = append(cidrs, ipNet)
+			}
+
+			var ipList []net.IP
+			for _, ip := range test.ipStrings {
+				parsedIP := net.ParseIP(ip)
+				if parsedIP == nil {
+					t.Fatalf("failed to parse IP %s", ip)
+				}
+				ipList = append(ipList, parsedIP)
+			}
+
+			cidrs = RemoveCIDRsContainingIPs(cidrs, ipList)
+			if !test.expectPanic {
+				if len(cidrs) != len(test.expectedCIDRs) {
+					t.Fatalf("unexpected number of remaining CIDRs: got %d, want %d", len(cidrs), len(test.expectedCIDRs))
+				}
+				for i, cidr := range cidrs {
+					if cidr.String() != test.expectedCIDRs[i] {
+						t.Errorf("unexpected CIDR at index %d: got %s, want %s", i, cidr.String(), test.expectedCIDRs[i])
+					}
+				}
+			}
+		})
+	}
+}
+
+func TestRemoveLargerOverlappingCIDRs(t *testing.T) {
+	type args struct {
+		cidrNets []*net.IPNet
+	}
+	tests := []struct {
+		name string
+		args args
+		want []*net.IPNet
+	}{
+		{
+			name: "equal",
+			args: args{
+				cidrNets: []*net.IPNet{
+					{IP: net.ParseIP("192.168.1.0"), Mask: net.CIDRMask(24, 32)},
+					{IP: net.ParseIP("192.168.2.0"), Mask: net.CIDRMask(24, 32)},
+				}},
+			want: []*net.IPNet{
+				{IP: net.ParseIP("192.168.1.0"), Mask: net.CIDRMask(24, 32)},
+				{IP: net.ParseIP("192.168.2.0"), Mask: net.CIDRMask(24, 32)},
+			},
+		},
+		{
+			name: "larger",
+			args: args{
+				cidrNets: []*net.IPNet{
+					{IP: net.ParseIP("192.168.1.0"), Mask: net.CIDRMask(24, 32)},
+					{IP: net.ParseIP("192.168.2.0"), Mask: net.CIDRMask(16, 32)},
+				}},
+			want: []*net.IPNet{
+				{IP: net.ParseIP("192.168.2.0"), Mask: net.CIDRMask(16, 32)},
+			},
+		},
+		{
+			name: "deduplicated",
+			args: args{
+				cidrNets: []*net.IPNet{
+					{IP: net.ParseIP("192.168.1.0"), Mask: net.CIDRMask(24, 32)},
+					{IP: net.ParseIP("192.168.1.0"), Mask: net.CIDRMask(24, 32)},
+				}},
+			want: []*net.IPNet{
+				{IP: net.ParseIP("192.168.1.0"), Mask: net.CIDRMask(24, 32)},
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := RemoveLargerOverlappingCIDRs(tt.args.cidrNets); !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("RemoveLargerOverlappingCIDRs() = %v, want %v", got, tt.want)
+			}
+		})
+	}
 }
